WOW - Typescript Porting is successful
This was an awesome exercises. Took help from Chat GPT 3.5.

Why did I port to Typescript?
The JSx build was working fine in simulator but did not work (launch) in android device

In the build log on expo server it showed some variables are not declared kinds of error. so it was not possible to find where it could have been problem.

So i am of the opinion TSX TS catch the error early type mismatch  so it should build fine because of no typescript errors